リファクタリング CTipWnd::ComputeWindowSize
ポインタ位置を示す変数の型を int → size_t

+++ sakura_core/window/CTipWnd.cpp
@@ -149,10 +149,10 @@ void CTipWnd::ComputeWindowSize(
 
 	int nCurMaxWidth = 0;
 	int nCurHeight = 0;
-	const int nTextLength = _tcslen( pszText );
-	for( int i = 0, nBgn = 0; i <= nTextLength; ++i ){
+	const size_t nTextLength = _tcslen( pszText );
+	for( size_t i = 0, nBgn = 0; i <= nTextLength; ++i ){
 		// 2005-09-02 D.S.Koba GetSizeOfChar
-		int nCharChars = CNativeT::GetSizeOfChar( pszText, nTextLength, i );
+		size_t nCharChars = CNativeT::GetSizeOfChar( pszText, nTextLength, i );
 		if( ( 1 == nCharChars && _T('\\') == pszText[i] && _T('n') == pszText[i + 1]) || _T('\0') == pszText[i] ){
 			RECT rc;
 			if( 0 < i - nBgn ){
